powerpc/timer - large decrementer support
POWER ISA v3 adds large decrementer mode of operation which increases
the size of the decrementer register from 32 bits to an implementation
defined with of up to 64 bits. This patch adds guest support for the LD
on processors with the CPU_FTR_ARCH_300 cpu feature flag set.

The current implementation enables LD mode when the timebase is being
initialised.  The boot processor will determines the decrementer size
from the devicetree using the ibm,dec-bits property which has been added
to the CPU nodes.

This patch was based on inital work by Jack Miller <millerjo@us.ibm.com>

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>