powerpc/mm/radix: Update LPCR HR bit as per ISA
The processor requires the MMU mode (radix vs. hash) of the hypervisor
to be mirrored in the LPCR register, in addition to the partition table,
so as to avoid fetching from the table when deciding, among other
things, how to perform transitions to HV mode on some interrupts. So
let's set it up appropriately

Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>